requirements.txt
colorgram.py==1.2.0
environ==1.0
elasticsearch==7.9.1
fastapi==0.61.1
filelock==3.0.12
Flask==1.1.2
Flask-Compress==1.6.0
future==0.18.2
gpg==1.13.1
h11==0.9.0
httptools==0.1.1
idna==2.10
importlib-metadata==1.3.0
ipykernel==5.1.3
ipython==7.10.2
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.10.3
jinja2-time==0.2.0
joblib==0.14.1
json5==0.8.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
jupyterlab==1.2.4
jupyterlab-server==1.0.6
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==2.2.3
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.0.2
nbconvert==5.6.1
nbformat==4.4.0
networkx==1.8.1
notebook==6.0.2
numpy==1.18.0
oauthlib==3.1.0
pandas==1.0.3
pandocfilters==1.4.2
parso==0.5.2
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.2.0
pipenv==2018.11.26
plotly-express==0.4.1
poyo==0.5.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
psycopg2==2.8.6
ptyprocess==0.6.0
pydantic==1.6.1
pygame==1.9.6
Pygments==2.5.2
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyparsing==2.4.5
pyrsistent==0.15.6
PySocks==1.7.1
python-dateutil==2.8.1
python-multipart==0.0.5
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2019.3
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
scikit-learn==0.22
scipy==1.4.1
seaborn==0.10.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.13.0
sklearn==0.0
soupsieve==2.0.1
sqlparse==0.3.0
starlette==0.13.6
statsmodels==0.12.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
tornado==6.0.3
traitlets==4.3.3
tweepy==3.9.0
typed-ast==1.4.0
urllib3==1.25.10
uvicorn==0.11.8
uvloop==0.14.0
virtualenv==20.0.23
virtualenv-clone==0.5.3
wcwidth==0.1.7
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
